+++ CHANGELOG.md
@@ -25,6 +25,9 @@
   [[#735](https://github.com/reupen/columns_ui/pull/735),
   [#740](https://github.com/reupen/columns_ui/pull/740)]
 
+- A built-in icon for the stop after current command was added to the Buttons
+  toolbar. [[#757](https://github.com/reupen/columns_ui/pull/757)]
+
 - When title formatting is used in the playlist switcher panel, typing in the
   panel now always searches by the actual playlist name and not the displayed
   title. [[#738](https://github.com/reupen/columns_ui/pull/738)]